requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-ckeditor==6.5.1
django-debug-toolbar==3.7.0
django-js-asset==2.0.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
psycopg2==2.8.6